fix: player rerender issue
